ai-test-gen/requirements.txt
anthropic>=0.39.0
PyGithub>=2.1.0
